refactor: domain 패키지 구조 세분화
