+++ src/main/java/domain/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.lotto;
 
 import java.util.ArrayList;
 import java.util.Collections;
