+++ src/main/java/domain/result/WinningResult.java
@@ -1,4 +1,7 @@
-package domain;
+package domain.result;
+
+import domain.lotto.Lotto;
+import domain.lotto.LottoBall;
 
 public class WinningResult {
     private final Lotto winningLotto;
