+++ src/test/java/domain/GameMoneyTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.money.GameMoney;
+import domain.lotto.LottoBundle;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
