+++ src/test/java/domain/LottoBallTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.lotto.LottoBall;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
