+++ src/test/java/domain/LottoResultTest.java
@@ -1,5 +1,11 @@
 package domain;
 
+import domain.lotto.Lotto;
+import domain.lotto.LottoBall;
+import domain.lotto.LottoBundle;
+import domain.result.LottoRank;
+import domain.result.LottoResult;
+import domain.result.WinningResult;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
