+++ src/test/java/domain/LottoTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.lotto.Lotto;
+import domain.lotto.LottoBall;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
